Change ID type from Int to Long and return it from Insert method

+++ app/src/main/java/com/soczuks/footballassistant/database/entities/CompetitionItem.kt
@@ -25,6 +25,6 @@ import androidx.room.Index
     indices = [Index(value = ["competitionId"]), Index(value = ["itemId"])]
 )
 class CompetitionItem(
-    val competitionId: Int,
-    val itemId: Int
+    val competitionId: Long,
+    val itemId: Long
 )
