+++ app/src/main/java/com/soczuks/footballassistant/ui/competitions/CompetitionsViewModel.kt
@@ -22,7 +22,7 @@ class CompetitionsViewModel(application: Application) : AndroidViewModel(applica
         competitionsDao.getAll()
     val viewModelMessage: StateFlow<ViewModelMessage?> = _viewModelMessage
 
-    fun getCompetitionById(id: Int): LiveData<CompetitionDetails>? {
+    fun getCompetitionById(id: Long): LiveData<CompetitionDetails>? {
         return competitionsDao.getById(id)
     }
